code_runner_service/requirements.txt
flask
docker
pika
python-dotenv
flask_sqlalchemy
flask_migrate
psycopg2-binary
psutil